[refactor] #49 등급 함수 utils/grade.ts 로 이동
